+++ MobicoolFR34.X/settings.c
@@ -13,7 +13,7 @@ void Settings_Initialize(settings_t* settings) {
     
     // Read settings
     settings->on = DATAEE_ReadByte(EE_ONOFF);
-    settings->temp_setpoint = DATAEE_ReadByte(EE_TEMP);
+    settings->temp_setpoint = (int8_t)DATAEE_ReadByte(EE_TEMP);
     if (settings->temp_setpoint < MIN_TEMP || settings->temp_setpoint > MAX_TEMP) {
         eeinvalid = true;
     }
@@ -29,13 +29,13 @@ void Settings_Initialize(settings_t* settings) {
         DATAEE_WriteByte(EE_ONOFF, settings->on);
         
         settings->temp_setpoint = DEFAULT_TEMP;
-        DATAEE_WriteByte(EE_TEMP, settings->temp_setpoint);
+        DATAEE_WriteByte(EE_TEMP, (uint8_t)settings->temp_setpoint);
         
         settings->fahrenheit = false;
         DATAEE_WriteByte(EE_UNIT, settings->fahrenheit);
         
         settings->battmon = BMON_LOW;
-        DATAEE_WriteByte(EE_BATTMON, settings->battmon);
+        DATAEE_WriteByte(EE_BATTMON, (uint8_t)settings->battmon);
         
         DATAEE_WriteByte(EE_MAGIC, MAGIC);
     }
@@ -46,7 +46,7 @@ void Settings_SaveOnOff(bool on) {
 }
 
 void Settings_SaveTemp(int8_t temp) {
-    DATAEE_WriteByte(EE_TEMP, temp);
+    DATAEE_WriteByte(EE_TEMP, (uint8_t)temp);
 }
 
 void Settings_SaveUnit(bool fahrenheit) {
@@ -54,5 +54,5 @@ void Settings_SaveUnit(bool fahrenheit) {
 }
 
 void Settings_SaveBattMon(bmon_t level) {
-    DATAEE_WriteByte(EE_BATTMON, level);
+    DATAEE_WriteByte(EE_BATTMON, (uint8_t)level);
 }
